moved PluginInfo and refactored parsing from properties file
