+++ core/src/main/java/org/elasticsearch/plugins/PluginsService.java
@@ -24,7 +24,6 @@ import com.google.common.collect.ImmutableMap;
 
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.Version;
-import org.elasticsearch.action.admin.cluster.node.info.PluginInfo;
 import org.elasticsearch.action.admin.cluster.node.info.PluginsInfo;
 import org.elasticsearch.bootstrap.Bootstrap;
 import org.elasticsearch.bootstrap.JarHell;
@@ -298,41 +297,6 @@ public class PluginsService extends AbstractComponent {
         List<PluginInfo> plugins = new ArrayList<>();
         List<URL> urls = new ArrayList<>();
     }
-    
-    /** reads (and validates) plugin metadata descriptor file */
-    static PluginInfo readMetadata(Path dir) throws IOException {
-        Path descriptor = dir.resolve(ES_PLUGIN_PROPERTIES);
-        Properties props = new Properties();
-        try (InputStream stream = Files.newInputStream(descriptor)) {
-            props.load(stream);
-        }
-        String name = dir.getFileName().toString();
-        String description = props.getProperty("description");
-        if (description == null) {
-            throw new IllegalArgumentException("Property [description] is missing for plugin [" + name + "]");
-        }
-        String version = props.getProperty("version");
-        if (version == null) {
-            throw new IllegalArgumentException("Property [version] is missing for plugin [" + name + "]");
-        }
-        String esVersionString = props.getProperty("elasticsearch.version");
-        if (esVersionString == null) {
-            throw new IllegalArgumentException("Property [elasticsearch.version] is missing for plugin [" + name + "]");
-        }
-        Version esVersion = Version.fromString(esVersionString);
-        if (esVersion.equals(Version.CURRENT) == false) {
-            throw new IllegalArgumentException("Elasticsearch version [" + esVersionString + "] is too old for plugin [" + name + "]");
-        }
-        boolean jvm = Boolean.parseBoolean(props.getProperty("jvm"));
-        boolean site = Boolean.parseBoolean(props.getProperty("site"));
-        boolean isolated = true;
-        String classname = "NA";
-        if (jvm) {
-            isolated = Boolean.parseBoolean(props.getProperty("isolated", "true"));
-            classname = props.getProperty("plugin");
-        }
-        return new PluginInfo(name, description, site, version, jvm, classname, isolated);
-    }
 
     static List<Bundle> getPluginBundles(Environment environment) throws IOException {
         ESLogger logger = Loggers.getLogger(Bootstrap.class);
@@ -350,7 +314,7 @@ public class PluginsService extends AbstractComponent {
             for (Path plugin : stream) {
                 try {
                     logger.trace("--- adding plugin [{}]", plugin.toAbsolutePath());
-                    PluginInfo info = readMetadata(plugin);
+                    PluginInfo info = PluginInfo.readFromProperties(plugin);
                     List<URL> urls = new ArrayList<>();
                     if (info.isJvm()) {
                         // a jvm plugin: gather urls for jar files
